Refactored scripts locations, added more docco Added IHittable for detecting hits.
