+++ Assets/NumbGoat/ProjectileSystems/Scripts.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d9943d006f2a9a74d8b1b1b741053a58
+folderAsset: yes
+timeCreated: 1505523288
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
