+++ Assets/NumbGoat/ProjectileSystems/Scripts/IMoving.cs
@@ -1,12 +1,12 @@
 ﻿using UnityEngine;
 
-namespace NumbGoat {
+namespace NumbGoat.ProjectileSystems.Scripts {
     /// <summary>
     ///     Interface for an object which moves.
     /// </summary>
     public interface IMoving {
         /// <summary>
-        /// Current velocity of this object.
+        ///     Current velocity of this object. Needed to aim at moving objects.
         /// </summary>
         Vector3 Velocity { get; }
     }
