+++ Assets/NumbGoat/ProjectileSystems/Scripts/Projectile/IHittable.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 7913256442f273645aa5cf5a884281eb
-timeCreated: 1504829236
+guid: e00980d3558cf074ba4c82574064814d
+timeCreated: 1505524450
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
