+++ Assets/NumbGoat/Scenes/MainTestScene.unity
@@ -598,6 +598,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   CurrentHealth: 100
   MaxHealth: 100
+  HealthRegen: 0.1
 --- !u!1 &596724160
 GameObject:
   m_ObjectHideFlags: 0
@@ -710,6 +711,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   CurrentHealth: 100
   MaxHealth: 100
+  HealthRegen: 0.1
 --- !u!1001 &697802545
 Prefab:
   m_ObjectHideFlags: 0
@@ -819,6 +821,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   CurrentHealth: 100
   MaxHealth: 100
+  HealthRegen: 0.1
 --- !u!54 &882118341
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -1377,7 +1380,7 @@ Transform:
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 1664286722}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 10, y: 0, z: -10}
+  m_LocalPosition: {x: 1.72, y: -8.85, z: -2.45}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 1859136692}
@@ -1396,6 +1399,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   CurrentHealth: 100
   MaxHealth: 100
+  HealthRegen: 0.1
 --- !u!1 &1789844023
 GameObject:
   m_ObjectHideFlags: 0
@@ -1734,7 +1738,7 @@ Transform:
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 1913659412}
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 14, y: 8, z: -5}
+  m_LocalPosition: {x: 20, y: 20, z: -5}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 1859136692}
@@ -1753,6 +1757,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   CurrentHealth: 100
   MaxHealth: 100
+  HealthRegen: 0.1
 --- !u!1 &1930846911
 GameObject:
   m_ObjectHideFlags: 0
@@ -1944,6 +1949,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   CurrentHealth: 100
   MaxHealth: 100
+  HealthRegen: 0.1
 --- !u!1 &2010948569
 GameObject:
   m_ObjectHideFlags: 0
@@ -2056,3 +2062,4 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   CurrentHealth: 100
   MaxHealth: 100
+  HealthRegen: 0.1
