+++ Assets/NumbGoat/Scripts/DefaultProjectile.cs
@@ -1,0 +1,11 @@
+﻿using NumbGoat.ProjectileSystems.Scripts.Projectile;
+using NumbGoat.Unit;
+using UnityEngine;
+
+namespace NumbGoat {
+    /// <summary>
+    ///     Default implementation of a projectile. Uses all of the properties in BaseProjectile for its logic.
+    /// </summary>
+    public class DefaultProjectile : BaseProjectile {
+    }
+}
