--- Assets/NumbGoat/Scripts/MovingHelpers.cs
@@ -1,3 +1,0 @@
-﻿namespace NumbGoat {
-    public static class MovingHelpers { }
-}
