+++ ProjectileSystems.csproj
@@ -12,12 +12,15 @@
     <ProjectTypeGuids>{E097FAD1-6243-4DAD-9C02-E9B9EFC3FFC1};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <TargetFrameworkIdentifier>.NETFramework</TargetFrameworkIdentifier>
     <TargetFrameworkVersion>v4.6</TargetFrameworkVersion>
-    <TargetFrameworkProfile></TargetFrameworkProfile>
-    <CompilerResponseFile></CompilerResponseFile>
+    <TargetFrameworkProfile>
+    </TargetFrameworkProfile>
+    <CompilerResponseFile>
+    </CompilerResponseFile>
     <UnityProjectType>Game:1</UnityProjectType>
     <UnityBuildTarget>StandaloneWindows64:19</UnityBuildTarget>
     <UnityVersion>2017.1.0p4</UnityVersion>
-    <RootNamespace></RootNamespace>
+    <RootNamespace>
+    </RootNamespace>
     <LangVersion>6</LangVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
@@ -74,19 +77,28 @@
     <Reference Include="UnityEngine.HoloLens">
       <HintPath>Library\UnityAssemblies\UnityEngine.HoloLens.dll</HintPath>
     </Reference>
+    <Reference Include="UnityEditor.iOS.Extensions.Xcode">
+      <HintPath>Library\UnityAssemblies\UnityEditor.iOS.Extensions.Xcode.dll</HintPath>
+    </Reference>
+    <Reference Include="UnityEditor.iOS.Extensions.Common">
+      <HintPath>Library\UnityAssemblies\UnityEditor.iOS.Extensions.Common.dll</HintPath>
+    </Reference>
+    <Reference Include="Mono.Cecil">
+      <HintPath>Library\UnityAssemblies\Mono.Cecil.dll</HintPath>
+    </Reference>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Assets\NavMeshComponents\Scripts\NavMeshLink.cs" />
     <Compile Include="Assets\NavMeshComponents\Scripts\NavMeshModifier.cs" />
     <Compile Include="Assets\NavMeshComponents\Scripts\NavMeshModifierVolume.cs" />
     <Compile Include="Assets\NavMeshComponents\Scripts\NavMeshSurface.cs" />
+    <Compile Include="Assets\NumbGoat\ProjectileSystems\Scripts\IMoving.cs" />
+    <Compile Include="Assets\NumbGoat\ProjectileSystems\Scripts\Projectile\BaseProjectile.cs" />
+    <Compile Include="Assets\NumbGoat\ProjectileSystems\Scripts\Projectile\IHittable.cs" />
+    <Compile Include="Assets\NumbGoat\Scripts\DefaultProjectile.cs" />
+    <Compile Include="Assets\NumbGoat\Scripts\ProjectileTestFire.cs" />
+    <Compile Include="Assets\NumbGoat\ProjectileSystems\Scripts\Projectile\TrajectoryHelper.cs" />
     <Compile Include="Assets\NumbGoat\Scripts\ExampleTarget.cs" />
-    <Compile Include="Assets\NumbGoat\Scripts\IMoving.cs" />
-    <Compile Include="Assets\NumbGoat\Scripts\MovingHelpers.cs" />
-    <Compile Include="Assets\NumbGoat\Scripts\Projectile\BaseProjectile.cs" />
-    <Compile Include="Assets\NumbGoat\Scripts\Projectile\DefaultProjectile.cs" />
-    <Compile Include="Assets\NumbGoat\Scripts\Projectile\ProjectileTestFire.cs" />
-    <Compile Include="Assets\NumbGoat\Scripts\Projectile\TrajectoryHelper.cs" />
     <Compile Include="Assets\NumbGoat\Scripts\Unit\WalkingUnit.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
